Benchmark limit

+++ tutorials/hybridCode/arr_sum.c
@@ -1,6 +1,6 @@
-int arr_sum(int numToSum)
+long int arr_sum(int numToSum)
 {
-    int squareSum = 0;
+    long int squareSum = 0;
     for (int i = 0; i < numToSum; i++)
     {
         squareSum += numToSum;
